Move error to later so we get it only if geo field is actually used in sort.
