+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/planner/LocalExecutionPlanner.java
@@ -337,17 +337,15 @@ public class LocalExecutionPlanner {
         List<Layout.ChannelSet> inverse = source.layout.inverse();
         for (int channel = 0; channel < inverse.size(); channel++) {
             elementTypes[channel] = PlannerUtils.toElementType(inverse.get(channel).type());
-            encoders[channel] = switch (inverse.get(channel).type().typeName()) {
+            String typeName = inverse.get(channel).type().typeName();
+            encoders[channel] = switch (typeName) {
                 case "ip" -> TopNEncoder.IP;
                 case "text", "keyword" -> TopNEncoder.UTF8;
                 case "version" -> TopNEncoder.VERSION;
                 case "boolean", "null", "byte", "short", "integer", "long", "double", "float", "half_float", "datetime", "date_period",
                     "time_duration", "object", "nested", "scaled_float", "unsigned_long", "_doc" -> TopNEncoder.DEFAULT_SORTABLE;
-                case "geo_point", "cartesian_point" -> throw new IllegalArgumentException(
-                    "can't sort on geo_point or point field without using specific sorting feature, like geo_distance"
-                );
-                case "geo_shape", "cartesian_shape" -> throw new IllegalArgumentException(
-                    "can't sort on geo_shape or shape field without using specific sorting feature, like geo_distance"
+                case "geo_point", "cartesian_point", "geo_shape", "cartesian_shape" -> TopNEncoder.unsortableWithError(
+                    "can't sort on " + typeName + " field"
                 );
                 // unsupported fields are encoded as BytesRef, we'll use the same encoder; all values should be null at this point
                 case "unsupported" -> TopNEncoder.UNSUPPORTED;
